core/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.3
django-filter==2.2.0
django-cors-headers==3.2.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
drf-extensions==0.6.0
idna==2.9
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pytz==2019.3
PyYAML==5.3
requests==2.23.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
psycopg2==2.8.5
